코콤 패킷에서 TYPE만 16비트로 두고, 장치·방·명령은 1바이트로 읽고 쓴다.

+++ Assets/Scripts/UI/KocomHexTestUI.cs
@@ -113,8 +113,8 @@ namespace Homepad.UI
         private bool linkUiApplying;
         private readonly Queue<byte[]> requestQueue = new Queue<byte[]>();
         private Coroutine requestBurstRoutine;
-        private ushort pendingAckDevice;
-        private ushort pendingAckRoom;
+        private byte pendingAckDevice;
+        private byte pendingAckRoom;
         private bool pendingAckSatisfied;
 
         private void Awake()
